Drop ex4 & add comment about llm install package in ex3 solution
