+++ w2d3/exercises/ex4/exercise/Dockerfile
@@ -2,11 +2,11 @@ FROM python:3.9-slim
 
 WORKDIR /app
 
-RUN apt-get update && apt-get install -y netcat-openbsd
-
 COPY src/requirements.txt .
+RUN pip install --no-cache-dir -r requirements.txt
 
 COPY src .
 
-COPY entrypoint.sh /entrypoint.sh
-ENTRYPOINT ["/entrypoint.sh"]
+EXPOSE 5000
+
+CMD ["python", "app.py"]
