--- w2d3/exercises/ex4/exercise/Dockerfile.developer
@@ -1,9 +1,0 @@
-FROM python:3.9-slim
-
-WORKDIR /app
-
-RUN pip install requests
-
-COPY src/developer_simulator.py .
-
-CMD ["python", "developer_simulator.py"]
