--- w2d3/exercises/ex4/exercise/Dockerfile.pypi
@@ -1,8 +1,0 @@
-FROM python:3.9-slim
-
-RUN pip install pypiserver['passlib']
-
-EXPOSE 8080
-COPY packages /packages
-
-CMD ["python", "-m", "pypiserver", "run", "-P", ".", "-a", "." ,"-p", "8080", "/packages"]
