+++ w2d3/exercises/ex4/exercise/docker-compose.yml
@@ -1,13 +1,4 @@
 services:
-  pypi-server:
-    build:
-      context: .
-      dockerfile: Dockerfile.pypi
-    ports:
-      - "8080:8080"
-    networks:
-      - workshop
-
   web:
     build:
       context: .
@@ -16,19 +7,5 @@ services:
       - "5000:5000"
     networks:
       - workshop
-    depends_on:
-      - pypi-server
-
-  developer-simulator:
-    build:
-      context: .
-      dockerfile: Dockerfile.developer
-    networks:
-      - workshop
-    depends_on:
-      - pypi-server
-      - web
-    command: [ "python", "developer_simulator.py" ]
-
 networks:
   workshop:
