--- w2d3/exercises/ex4/exercise/src/external-requirements.txt
@@ -1,3 +1,0 @@
-flask==2.3.3
-requests==2.31.0
-
