--- w2d3/exercises/ex4/exercise/src/flower-power/MANIFEST.in
@@ -1,2 +1,0 @@
-recursive-include flower_power/images *
-
