--- w2d3/exercises/ex4/exercise/src/internal-requirements.txt
@@ -1,2 +1,0 @@
-flower_power==2.0.0
-
