+++ w2d3/exercises/ex4/exercise/src/leaderboard.json
@@ -193,30 +193,5 @@
     "player_name": "Nitzan",
     "apple_size": 1337,
     "timestamp": "2025-08-10T21:09:21.931283"
-  },
-  {
-    "player_name": "Playeraa",
-    "apple_size": 18,
-    "timestamp": "2025-08-12T11:11:07.402391"
-  },
-  {
-    "player_name": "Playeraa",
-    "apple_size": 20,
-    "timestamp": "2025-08-12T11:11:19.877143"
-  },
-  {
-    "player_name": "Playeraa",
-    "apple_size": 20,
-    "timestamp": "2025-08-12T11:11:28.595593"
-  },
-  {
-    "player_name": "nitzan",
-    "apple_size": 10005,
-    "timestamp": "2025-08-12T12:48:09.689332"
-  },
-  {
-    "player_name": "\"aWhXJ<f}is%/M#qvS\";p>()#tt(<)Oni)co?|_Pa)-{atRkzm]=z)!Xrwb0R/0vhx+~?(70SYB+IemMAhHpW9j$m*f q)Q}:TW,zg{*SfSsumAR,@SX  $ )   u?!s",
-    "apple_size": 2,
-    "timestamp": "2025-08-12T14:08:13.357989"
   }
 ]
