+++ w2d3/exercises/ex4/exercise/src/requirements.txt
@@ -1,6 +1,2 @@
-# This file aggregates both external and internal requirements.
-# External deps are installed at build time from PyPI.
-# Internal deps (like flower_power) are installed at runtime from the local PyPI server.
-
--r external-requirements.txt
--r internal-requirements.txt
+Flask==2.3.3
+requests==2.31.0
