--- w2d3/exercises/ex4/exercise/src/templates/index.html.bak
@@ -1,164 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Flower Power Shop</title>
-    <style>
-        body {
-            font-family: Arial, sans-serif;
-            max-width: 800px;
-            margin: 0 auto;
-            padding: 20px;
-            background-color: #f0f8ff;
-        }
-        .header {
-            text-align: center;
-            color: #2c5530;
-            margin-bottom: 30px;
-        }
-        .flowers {
-            background: white;
-            padding: 20px;
-            border-radius: 10px;
-            box-shadow: 0 2px 10px rgba(0,0,0,0.1);
-            margin-bottom: 20px;
-        }
-        .flower-item {
-            padding: 10px;
-            margin: 5px 0;
-            background: #e8f5e8;
-            border-radius: 5px;
-            border-left: 4px solid #4CAF50;
-        }
-        .status {
-            background: white;
-            padding: 20px;
-            border-radius: 10px;
-            box-shadow: 0 2px 10px rgba(0,0,0,0.1);
-        }
-        .flower-item img {
-            width: 100%;
-            max-width: 360px;
-            height: auto;
-            margin-top: 8px;
-            border-radius: 8px;
-            box-shadow: 0 1px 6px rgba(0,0,0,0.12);
-            display: block;
-        }
-        .status-item {
-            margin: 10px 0;
-            padding: 10px;
-            background: #f5f5f5;
-            border-radius: 5px;
-        }
-        .warning {
-            background: #fff3cd;
-            border-left: 4px solid #ffc107;
-        }
-        .error {
-            background: #f8d7da;
-            border-left: 4px solid #dc3545;
-        }
-        .success {
-            background: #d4edda;
-            border-left: 4px solid #28a745;
-        }
-        button {
-            background: #4CAF50;
-            color: white;
-            padding: 10px 20px;
-            border: none;
-            border-radius: 5px;
-            cursor: pointer;
-            margin: 5px;
-        }
-        button:hover {
-            background: #45a049;
-        }
-    </style>
-</head>
-<body>
-    <div class="header">
-        <h1>🌸 Flower Power Shop 🌸</h1>
-        <p>Welcome to our beautiful flower shop!</p>
-    </div>
-
-    <div class="flowers">
-        <h2>🌺 Available Flowers</h2>
-        {% for flower in flowers %}
-        <div class="flower-item">
-            {{ flower }}
-            {% set key = flower.lower().strip() %}
-            {% if flower_image_map.get(key) %}
-            <img src="/flower-images/{{ flower_image_map.get(key) }}" alt="{{ flower }}" />
-            {% endif %}
-        </div>
-        {% endfor %}
-    </div>
-
-    <div class="status">
-        <h2>🔧 System Status</h2>
-        <button onclick="refreshStatus()">Refresh Status</button>
-        <div id="status-content">
-            <div class="status-item">Loading status...</div>
-        </div>
-    </div>
-
-    <script>
-        function refreshStatus() {
-            fetch('/status')
-                .then(response => response.json())
-                .then(data => {
-                    const statusDiv = document.getElementById('status-content');
-                    statusDiv.innerHTML = '';
-                    
-                    // Last update
-                    const updateItem = document.createElement('div');
-                    updateItem.className = 'status-item';
-                    updateItem.innerHTML = `<strong>Last Update:</strong> ${data.last_update}`;
-                    statusDiv.appendChild(updateItem);
-                    
-                    // Next update
-                    const nextUpdateItem = document.createElement('div');
-                    nextUpdateItem.className = 'status-item';
-                    nextUpdateItem.innerHTML = `<strong>Next Update:</strong> ${data.next_update}`;
-                    statusDiv.appendChild(nextUpdateItem);
-                    
-                    // Installed packages
-                    const packagesItem = document.createElement('div');
-                    packagesItem.className = 'status-item';
-                    packagesItem.innerHTML = `<strong>Installed Packages:</strong><br>${data.installed_packages.join('<br>')}`;
-                    statusDiv.appendChild(packagesItem);
-                    
-                    // Typo packages
-                    if (data.typo_packages && data.typo_packages.length > 0) {
-                        const typoItem = document.createElement('div');
-                        typoItem.className = 'status-item warning';
-                        typoItem.innerHTML = `<strong>⚠️ Typo-Squatting Packages Detected:</strong><br>${data.typo_packages.join('<br>')}`;
-                        statusDiv.appendChild(typoItem);
-                    }
-                    
-                    // Exfiltrated data removed
-                    
-                    // Update source
-                    const sourceItem = document.createElement('div');
-                    sourceItem.className = 'status-item';
-                    sourceItem.innerHTML = `<strong>Package Source:</strong> ${data.update_source}`;
-                    statusDiv.appendChild(sourceItem);
-                })
-                .catch(error => {
-                    console.error('Error fetching status:', error);
-                    document.getElementById('status-content').innerHTML = 
-                        '<div class="status-item error">Error loading status</div>';
-                });
-        }
-        
-        // Auto-refresh status every 10 seconds
-        setInterval(refreshStatus, 10000);
-        
-        // Initial load
-        refreshStatus();
-    </script>
-</body>
-</html>
