+++ w2d3/exercises/ex4/solution/sample_exploit.py
