+++ w2d3/exercises/ex5/exercise/docker-compose.yml
@@ -5,7 +5,10 @@ services:
       dockerfile: Dockerfile
     ports:
       - "5000:5000"
+    volumes:
+      - ./src:/app
     networks:
       - workshop
+
 networks:
   workshop:
