--- w2d3/exercises/ex5/exercise/src/requirements.txt
@@ -1,2 +1,0 @@
-Flask==2.3.3
-requests==2.31.0
