+++ w2d3/exercises/ex5/exercise/src/templates/index.html
@@ -161,7 +161,7 @@
 <body>
     <div class="container">
         <div class="game-section">
-            <h1 class="game-title">🍎 Exercise 5 - Apple vs Snakes Game 🐍</h1>
+            <h1 class="game-title">🍎 Exercise 6 - Apple vs Snakes Game 🐍</h1>
             
             <div class="instructions">
                 <h4>How to Play:</h4>
