--- w2d3/exercises/ex6/exercise/Dockerfile
@@ -1,12 +1,0 @@
-FROM python:3.9-slim
-
-WORKDIR /app
-
-COPY requirements.txt .
-RUN pip install --no-cache-dir -r requirements.txt
-
-COPY src/ .
-
-EXPOSE 5000
-
-CMD ["python", "app.py"]
