--- w2d3/exercises/ex6/exercise/docker-compose.yml
@@ -1,14 +1,0 @@
-services:
-  web:
-    build:
-      context: .
-      dockerfile: Dockerfile
-    ports:
-      - "5000:5000"
-    volumes:
-      - ./src:/app
-    networks:
-      - workshop
-
-networks:
-  workshop:
